Apply suggestions from code review

+++ esphome/components/sonoff_d1/sonoff_d1.cpp
@@ -67,7 +67,7 @@ void SonoffD1Output::populate_checksum_(uint8_t *cmd, const size_t len) {
 }
 
 void SonoffD1Output::skip_command_() {
-  unsigned int garbage = 0;
+  size_t garbage = 0;
   // Read out everything from the UART FIFO
   while (this->available()) {
     uint8_t value = this->read();
@@ -131,7 +131,7 @@ bool SonoffD1Output::read_ack_(const uint8_t *cmd, const size_t len) {
   // Expected acknowledgement from rf chip
   uint8_t ref_buffer[7] = {0xAA, 0x55, cmd[2], cmd[3], 0x00, 0x00, 0x00};
   uint8_t buffer[sizeof(ref_buffer)] = {0};
-  unsigned int pos = 0, buf_len = sizeof(ref_buffer);
+  uint32_t pos = 0, buf_len = sizeof(ref_buffer);
 
   // Update the reference checksum
   this->populate_checksum_(ref_buffer, sizeof(ref_buffer));
